+++ tests/test_quick_blueprint.py
@@ -15,12 +15,20 @@ client_version = '0.9.5'
 def test_quick_blueprint():
     with Client(app) as client:
 
+        realProjectFiles = ['src/extension.ts',
+                            'src/test/runTest.ts',
+                            'src/test/suite/index.ts',
+                            'src/test/suite/extension.test.ts',
+                            'src/test/suite/extension.test.ts',
+                            'package.json']
+        projectFileList = realProjectFiles.copy()
+        likelyExclusionList = ['src/info.txt',
+                               'src/changelog.md',
+                               'src/sample.html',
+                               'src/branding.jpg']
+        projectFileList.extend(likelyExclusionList)
         request_body = {
-            'filelist': ['src/extension.ts',
-                         'src/test/runTest.ts',
-                         'src/test/suite/index.ts',
-                         'src/test/suite/extension.test.ts',
-                         'package.json'],
+            'filelist': projectFileList,
             'projectName': 'typescript-sample-extension',
             'session': 'testemail: unittest@polytest.ai',
             'organization': 'polytest.ai',
@@ -42,7 +50,21 @@ def test_quick_blueprint():
         assert 'draftBlueprint' in details
         assert 'recommendedSampleSourceFile' in details
         assert 'recommendedProjectDeploymentFile' in details
+        assert 'recommendedListOfFilesToExcludeFromAnalysis' in details
 
+        # we want high confidence exclusions only
+        excludedAtLeastOne = False
+        for file in likelyExclusionList:
+            if file in details['recommendedListOfFilesToExcludeFromAnalysis']:
+                excludedAtLeastOne = True
+                break
+        assert excludedAtLeastOne
+
+        # but we don't want to exclude real source ever
+        for file in realProjectFiles:
+            assert file not in details['recommendedListOfFilesToExcludeFromAnalysis']
+
+        request_body['projectFileList'] = list(set(projectFileList) - set(details['recommendedListOfFilesToExcludeFromAnalysis']))
         request_body['projectFile'] = package_json
         request_body['draftBlueprint'] = details['draftBlueprint']
         request_body['code'] = sample_code_ts